+++ lib/utils/multiPriceListElement.dart
@@ -1,6 +1,6 @@
 class MultiPriceListElement {
-  final int unitsFrom;
-  final int unitsTo;
+  final double unitsFrom;
+  final double unitsTo;
   final double price;
   MultiPriceListElement (this.unitsFrom, this.unitsTo, this.price);
 }
